Add an annotated introduction to Ruby
This takes the code examples from Chapter 1, produced in the book in use
with irb, and puts them into a .rb file with test cases asserting the
output that is show in the book.

It's not very exciting, really it's just a basic introduction to the
Ruby language. The following chapters will take up the mandate of the
book and give us the exciting things.