Create Simple Lang + VM to run it in
Using small-step semantics we create a language with a few basic
operations, so far we have Numbers, Booleans, +, *, less-than, variable
assignment and reading, if statements and while loops.

The whole thing is run inside a VM that takes the statements and
expressions and reduces them down until they can be reduced no
more. Thus producing a DoNothing (end state) for statements and a value
output for expressions.

Unfortunately the differences between expressions and statements are
starting to get in the way here. For example, an expression cannot
contain statements, e.g. assigning a variable, but a statement cannot
return any result other than the fully reduced DoNothing.